+++ Peak.StableThirdPersonCamera/Cinemachine.cs
@@ -1,5 +1,0 @@
-using Unity.Cinemachine;
-using UnityEngine;
-
-namespace Linkoid.Peak.StableCamera;
-